separate better the collections matchers
